sprint_8-solution-time-and-duration. Добавить время начала и продолжительность задачи. Написать и разделить тесты. Добавить список приоритезации по времени начала. Добавить проверку по пересечению задач.
