+++ src/Main.java
@@ -4,13 +4,16 @@ import model.Epic;
 import model.SubTask;
 import model.Task;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
+
 public class Main {
 
     public static void main(String[] args) {
         System.out.println("Поехали!");
         InMemoryTaskManager inMemoryTaskManager = new InMemoryTaskManager();
-        Task task1 = new Task("Задача 1", "Задача_1", Status.NEW);
-        Task task2 = new Task("Задача 2", "Задача_2", Status.NEW);
+        Task task1 = new Task("Задача 1", "Задача_1", Status.NEW, LocalDateTime.now(), Duration.ofMinutes(1));
+        Task task2 = new Task("Задача 2", "Задача_2", Status.NEW, LocalDateTime.now(), Duration.ofMinutes(1));
 
         inMemoryTaskManager.addTask(task1);
         inMemoryTaskManager.addTask(task2);
@@ -22,9 +25,9 @@ public class Main {
         inMemoryTaskManager.addEpic(epic1);
         inMemoryTaskManager.addEpic(epic2);
 
-        SubTask subTask1 = new SubTask("Подзадача 1", "Подзадача_1", epic1.getId(), Status.NEW);
-        SubTask subTask2 = new SubTask("Подзадача 2", "Подзадача_2", epic2.getId(), Status.NEW);
-        SubTask subTask3 = new SubTask("Подзадача 3", "Подзадача_3", epic2.getId(), Status.NEW);
+        SubTask subTask1 = new SubTask("Подзадача 1", "Подзадача_1", epic1.getId(), Status.NEW, LocalDateTime.now(), Duration.ofMinutes(1));
+        SubTask subTask2 = new SubTask("Подзадача 2", "Подзадача_2", epic2.getId(), Status.NEW, LocalDateTime.now(), Duration.ofMinutes(1));
+        SubTask subTask3 = new SubTask("Подзадача 3", "Подзадача_3", epic2.getId(), Status.NEW, LocalDateTime.now(), Duration.ofMinutes(1));
 
         inMemoryTaskManager.addSubTask(subTask1);
         inMemoryTaskManager.addSubTask(subTask2);
