+++ src/exception/ManagerValidateException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class ManagerValidateException extends RuntimeException{
+    public ManagerValidateException(String message) {
+        super(message);
+    }
+}
