+++ test/manager/InMemoryHistoryManagerTest.java
@@ -1,6 +1,7 @@
 package manager;
 
 import enumeration.Status;
+import exception.ManagerValidateException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -8,6 +9,8 @@ import model.Epic;
 import model.SubTask;
 import model.Task;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.List;
 
 class InMemoryHistoryManagerTest {
@@ -21,7 +24,7 @@ class InMemoryHistoryManagerTest {
 
     @Test
     void addTask() {
-        Task task = new Task("Test task", "Test task description", Status.NEW);
+        Task task = new Task("Test task", "Test task description", Status.NEW, LocalDateTime.now(), Duration.ofMinutes(5));
         taskManager.addTask(task);
         Epic epic = new Epic("Test epic", "Test epic description");
         taskManager.addEpic(epic);
@@ -29,7 +32,9 @@ class InMemoryHistoryManagerTest {
                 "Test subTask1",
                 "Test subTask1 description",
                 epic.getId(),
-                Status.IN_PROGRESS);
+                Status.IN_PROGRESS,
+                LocalDateTime.now().minusHours(1),
+                Duration.ofMinutes(5));
         taskManager.addSubTask(subTask);
         List<Task> history = taskManager.getHistory();
         Assertions.assertEquals(0, history.size(), "Количество задач в истории не равно нулю");
